mke2fs.conf: adjust small.inode_ratio to 16384, fix userdata.img resize issue

+++ tools/update_prebuilt.sh
@@ -11,12 +11,21 @@ cp -f $2/idbloader.img $1/
 cp -f $2/misc.img $1/
 cp -f $2/dtbo.img $1/
 
+TOP=$PWD
+export MKE2FS_CONFIG="${TOP}/tools/mke2fs.conf"
+if [ ! -f ${MKE2FS_CONFIG} ]; then
+    echo "error: ${MKE2FS_CONFIG} not found."
+    exit 1
+fi
+true ${MKFS:="${TOP}/tools/mke2fs"}
+
 RET=0
 if [ ! -f $1/userdata.img ]; then
 	USERDATA_SIZE=209715200
 	echo "Generating empty userdata.img (size:${USERDATA_SIZE})"
 	TMPDIR=`mktemp -d`
-	${PWD}/tools/make_ext4fs -s -l ${USERDATA_SIZE} -a root -L userdata $1/userdata.img ${TMPDIR}
+	IMG_BLK=$((${USERDATA_SIZE} / 4096))
+	${MKFS} -E android_sparse -t ext4 -L userdata -M /userdata -b 4096 -d ${TMPDIR} $1/userdata.img ${IMG_BLK}
 	RET=$?
 	rm -rf ${TMPDIR}
 fi
